Cornerstone Global Elements
I’ve searched for a couple possibilities now to get the stuff working.
The good thing is: I have already built an environment to get the elements work. Unfortunately the styles are still missing. 
Feel free to extend the code as you think it fits the best.
Would love if you have some ideas about getting the css and js stuff working.

Necessary steps:
1. Setup the postdata of your global element and fire the template_redirect hook again for setting up the flattened cs_element_shortcode_data
The rest is quite magical, so just copy it from the file

For this example I have quickly copied the function global_blocks_output for from the following file:
wp-content/plugins/global-blocks-for-cornerstone/element.php in line 51

Cheers
Jannis